Refactor plotting utilities into a modular package
- Split apts/plotting/utils.py into apts/plotting/utils/ package.
- Created geometry.py, style.py, and marking.py sub-modules.
- Refactored mark_observation to reduce cyclomatic complexity.
- Maintained public API via __init__.py re-exports.
- Deleted the original apts/plotting/utils.py file to avoid shadowing.
